Promote FlatIO to package chisel3
Also change to an object (from a def) so that it shows up in ScalaDoc
search.
